Add script with default seeding values
